Revert "Fixed a crash when accessing the bounds property of a UIView"
This reverts commit 478dacd22f652a1c33872f3f837e614436254f59.

+++ Frameworks/StarboardXaml/XamlCompositor.cpp
@@ -297,7 +297,7 @@ float DisplayNode::GetPresentationPropertyValue(const char* name) {
     std::string str(name);
     std::wstring wstr(str.begin(), str.end());
 
-    return (float)xamlNode->Get(ref new Platform::String(wstr.data()));
+    return (float)(double)xamlNode->Get(ref new Platform::String(wstr.data()));
 }
 
 void DisplayNode::SetContentsCenter(float x, float y, float width, float height) {
